interpreter: method calls, control flow, ref-counted objects
This is a bit of a patch-bomb but I think it's justifiable given how
immature the current implementation is that small incremental patches
would be too much of a pain.

Anyway, this adds support for parsing LocalX/ArgX as well as method
calls, code control flow like If/Else/While. The patch also makes the
interpreter respect the new ref-counted object model.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>