Make is_rev1 a global
Previously we would store it per-method and every new method would
inherit the revision from parent. Even though this made sense this is
not what the NT implementation does, instead it records the DSDT
revision and treats all other definition blocks as that same revision,
so we must do it that way as well.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>