Rewrite the interpreter/parser to be a micro-per-op-vm
This makes the intepreter a micro VM that breaks down every AML opcode
into smaller sub-opcodes that it operates on.

The change allows us to focus on writing actual opcode logic, abstracted
away from ad-hoc parsing of the AML bytestream. This also greately
reduces code duplication as all of the sub-ops are shared with only a
few of them being opcode-specific.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>