interpreter: fix Store(..., <ref>) semantics
Previously, storing an object to a reference to an object of the same
type would resize: e.g. storing "hello" to "xyz" would overwrite the
value and result in "hello". What should actually be taking place here
is a regular "implicit cast", so the correct resuling value is "hel" aka
"hello" truncated to the size of the destination buffer.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>