Make CopyObject & non-ref Store actually replace the object
Previously we would overwrite the target object instead of replacing the
object pointer at the parent object, which would lead to all sorts of
unnecessary workarounds and potential problems down the line when we
start implementing Package & Index. With this change, it should now be
much easier as we have the invariant that object type is immutable.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>