interpreter: TERM_ARG_NO_INVOKE -> TERM_ARG_OR_NAMED_OBJECT
No invoke used to imply don't convert named objects that are methods to
method calls, and that was fine because that was all that we supported.

But ACPI spec defines other object types besides methods that have the
magical behavior of automatically doing something when referenced, for
example buffer field or unit field objects.

Rename this parse op and clarify what it does so that we can reuse it
for field objects in the future commits.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>