Introduce the idea of indirect/out-of-line ops
Some of the more complex instructions take more than 16 uop-opcodes to
implement the decoding of, so we have to have a way to feed the parser
bigger op buffers. We do this by introducing the new op property called
UACPI_OP_PROPERTY_OUT_OF_LINE. This property is set for all ops that
have separate opcode buffers that don't fit into the inline 16-byte
buffer.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>