Rework package parsing for named objects
When we encounter a named object while parsing a package we shouldn't
actually be trying to resolve the name, but instead convert it to a path
string and store that instead to be resolved later when actually needed.

The reason for that is because often times a named package element
actually references a named object defined later on in the AML.

Acpi.sys decided to take the simple approach and do what I've described
in the first paragraph. So when a package uses a named object as one of
its elements, even if it has actually been defined and can be resolved
it still isn't resolved (ever). So the AML code can never actually
interact with the named object itself via the package because as far as
the package is concerned it's just a string.

ACPICA decided to work around this by introducing a second pass to its
interpreter, where package elements are resolved after parsing the
entire namespace. This works, however, is a bit of an overkill.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>