test-cases: introduce a test case for copying control methods
This is one of the crazier tests that don't work on any implementation
I've tried. (A simplified version of this works on NT, if GETX is not
marked as Serialized and is not a local object). All other existing
implementations either crash or produce a use-after-free or other
similar errors.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>