Greately simplify the table subsystem
Completely get rid of the reference-count system for tables and always
keep them mapped.

Tables are typically very small in size, perhaps on huge servers they
could go up as far as multiple megabytes, which is basically nothing for
a 64-bit address space. For 32-bit computers I don't expect that tables
could get very big, but even if they do, it's still pretty much
nothing.

The system prior to this commit is also very fragile: when we initialize
the namespace it will inevitably contain a bunch of control methods, and
every control method has a pointer, which is an offset into a virtual
mapping of the respective table the method was created in. Therefore we
simply cannot afford to unmap the DSDT/SSDT/etc. tables anyway, which
are usually the biggest tables anyway.

For now just get rid of the ability to unmap tables at runtime and maybe
reintroduce later when the need arises.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>