tables: introduce table_identifiers and rework table search
This reworks table search to use a public uacpi_table_identifiers
structure, which consists of a signature as well as oem_id and
oem_table_id strings. The find_by_type helper is reworked so that for
dynamic table types it converts the type to a signature string and
searches that instead for uniformity.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>