Implement Load
A thing of note for the future here:
The Buffer/OpRegion parsing is implemented as TERM_ARG_UNWRAP_INTERNAL
because one possible valid argument for the Load instruction is a field
unit, and we want it to be actually read beforehand, aka not take it in
as a SuperName, but as an integer or (hopefully) a Buffer object.

However, the specification tells us to expect a NameString, which means
the current implementation is technically wrong in the sense that, for
example, if we see a named method object we will try to invoke it, which
is wrong. This is not really that big of an issue simply because a
method object is not a valid argument for Load. In the end this will
result in us printing the wrong error message or something along these
lines. I don't consider this a problem, but it's worth noting.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>