Factor out uacpi_load_tabe into a separate helper
So that we can share this code between the initial table load code
as well as the dynamic AML Load/LoadTable operators. This also
adds a new table_load_cause parameter to the do_load_table helper in
the interpreter code. This parameter is used to generate the
appropriate log message depending on whether this is a static or a
dynamic table load.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>