uacpi: introduce uacpi_eval_typed
This also introduce UACPI_STATUS_TYPE_MISMATCH as well as object type
bit values to use for the ret_mask of the new helper.

The helper will be used to evaluating methods where the expected return
value is known beforehand so that the caller has to do less work of
actually validating it.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>