Implement _OSI
This adds bare bones support for _OSI that has a hardcoded list of
supported features.

An obvious TODO to make this dynamic in the future.

The fact that it had to be implemented in interpreter.c is also a FIXME
of course, but this is currently the only way that could be done because
all of the interpreter structures and helpers are private & local.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>