Introduce UACPI_STATUS_AML_UNDEFINED_REFERENCE
This introduces the very first UACPI_STATUS_AML_* error code, with
hopefully more of these coming later. This will be used to differentiate
between errors that have internal library origin as opposed to
bytecode-related errors.

This one specifically has been a bit of a pain point because without it
we were not able to distinguish when a uacpi_eval failed because the
requested path did not exist as opposed to bytecode referencing an
object that does not exist.

In the future I'm hoping to get rid of UACPI_STATUS_BAD_BYTECODE
entirely in favor of more specific AML error codes.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>