Introduce UACPI_STATUS_AML_BAD_ENCODING
This error code is used to report all errors related to bad instruction
encoding, for example when a string is not null-terminated, or a buffer
declares less objects that indicated by the package length, etc.

This commit also removes UACPI_STATUS_OUT_OF_BOUNDS as it is no longer
needed nor used for anything.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>