Introduce notify handlers & utility objects
This introduces the base objects that will be used to implement Notify()
callbacks. uacpi_handlers had to be renamed to
uacpi_address_space_handlers since an operation region object cannot be
used in a Notify call and therefore will not have a notify handler list
as part of the object.

uacpi_handlers is now a generic struct that contains both address space
as well as notify handlers and is a part of all device-like objects.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>