Implement proper FADT parsing & unified format conversion
FADT is notorious for having many versions that a lot of firmware
implementations get confused about misrepresenting their version or
storing garbage in reserved fields. There's also the problem of dealing
with legacy vs x_* fields.

This commit adds proper handling to all of those edge cases and converts
FADT to a unified format where we can always safely access the x_*
fields, as well as corrects the revision advertised by the table.

Fixes: #42
Signed-off-by: Daniil Tatianin <99danilt@gmail.com>