Add support for fields & buffer indices + write-back for inc/dec
Previously we would segfault when attempting to write to a buffer index
as it was being treated as a reference. We would also completely ignore
write-back for field units so trying to do FILD++ would not actually
update the field object itself, only the temporary integer value
read from the field.

Fixes: #45
Signed-off-by: Daniil Tatianin <99danilt@gmail.com>