uacpi: correct logic for choosing RSDT vs XSDT
Previously we would check for revision > 0, which is actually wrong
We're instead supposed to be checking for revision > 1 because that's
the revision when the XSDT field was added.

This also adds an additional check to make sure that XSDT is actually
not NULL before grabbing it just to be safe.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>