Introduce reduced-hardware-only mode
This mode strips out any ACPI-hardware support out of the build for
platforms that aren't expected to have that implemented, e.g. ARM.

In order to accomodate this, a new status COMPILED_OUT is added to
signify to the caller that the requestion functionality is not included
in the current build.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>