utilities: introduce uacpi_find_devices_at & make hid list terminated
The new helper allows specifying the scope of where to start searching
as well as the ability to provide multiple HIDs.

The HID list is now terminated by a UACPI_NULL entry instead of
specifying the entry count everywhere. This seemed more intuitive and
less error prone, while also allowing to simplify the code quite a bit.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>