Support skipping OPs for undefined/duplicate objects during a ns load
There are AML blobs in the wild that attempt to create duplicate named
objects or reference names that don't exist due to various BIOS
configuration errors during a table load (often times for SSDTs).
Aborting the namespace load in that case is not ideal as we can cleanly
skip a lot of the opcodes that might try do this.

This patch introduces a new "compressed" opcode called
SKIP_WITH_WARN_IF_NULL, that reduces the previous IF_NULL,
SKIP_WITH_WARN combination by two bytes.

Most of the existing opcodes are converted to use it, some exceptions
are Field objects as it's unclear what the best course of action would
be if we encounter any duplicate field units. For now just abort on
those.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>