utilities: introduce uacpi_string_to_integer
This introduces a new helper for converting strings to numbers.

Previously we would depend on a host-provided strtoull implementation
for this, but this is a not a commonly seen function that kernels
implement and its accuracy is crucial for making sure GPE matching and
AML ToInteger operator work correctly.

Because of this, let's ship our own independent implementation of string
conversions that we have verified to do what we want at a decent
performance level.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>