tests/runner: make physical memory emulation a bit more advanced
This makes us actually track the physical addresses we have "mapped"
previously, and return them instead of allocating an entirely new
buffer. Reference counting for mappings is added to make this work for
multiple users. This is important to be able to write to operation
regions from AML and then read back the same content when doing a table
load.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>