utilities: rewrite uacpi_eval_{hid,cid} to do a single allocation
Previously uacpi_pnp_id_list was allocated as an array of strings where
each string was a separate allocation, this heavily increased malloc
traffic and was hard to deal with for frees.

Rewrite both to produce a single allocation instead. This also greately
simplifies freeing and the transition to sized frees in the following
commits.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>