Introduce a UACPI_SIZED_FREES mode
This mode (when enabled) changes uacpi_kernel_free to take in a second
'size_hint' parameter, which contains the original size that was passed
to uacpi_kernel_alloc/calloc for allocators that can make use of this
information either for optimizations or by necessity.

To facilitate this, all uACPI frees must make use of uacpi_free as
opposed to directly calling uacpi_kernel_free to make sure they work in
both modes.

This also reworks the test runner to support this mode and track all
allocations in a hash table for verifying 'size_hint' correctness.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>