interpreter: make method aborts non-fatal & simplify unwind code
This makes it so encountering an error while in a method call is no
longer fatal if doing a table load. Instead, the stack is unwound back
to table level code, which resumes execution.

As an added bonus, this simplifies the stack unwind code quite a bit
because we remove an unneded unwind call from execution_context_release.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>