Introduce loop timeouts
Add a configurable timeout value in seconds after which a currently
running While loop is aborted to prevent unguarded infinite loops that
a lot of AML code contains.

To facilitate this, a new UACPI_STATUS_AML_LOOP_TIMEOUT code is
introduced and is returned if a loop is aborted.

This also adds a public context header that offers helpers to override
the loop timeout value if desired.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>