Introduce helpers for mutex management & optional tracing
This adds helpers for acquiring and releasing mutexes that should be
used over direct calls to uacpi_kernel_acqire/release. The reason they
are better is because they offer optional tracing of acquires and
releases, which can be used to debug potential deadlocks, as well as
handling common code duplication wherein an infinite timeout mutex
acquire needs to be checked at every callsite in case it returns false.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>