Rework the table subsystem to be thread safe & extensible
The table subsystem was probably the weakest and least featureful link
in the entire project. This rewrite aims to bring it up to part with the
rest of the library. In particular, this adds the ability to override,
deny, and otherwise control the tables that are being loaded at runtime.
This also adds public helpers for dynamically loading or installing
tables by the host.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>