tables: introduce reference counting
Make tables reference counted (again). This time, we do this as a
stepping stone to supporting early (heapless) table access, which needs
reference counting in order to enforce early mapping semantics, as well
as to prevent wasting of virtual space for the early virtual memory
management subsystem of a kernel, which might have a very limited number
of page slots.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>