interpreter: unconditionally bump the reference count of nodes
We have to do this to allow references to temporary nodes that might be
uninstalled from a different AML thread while we block on something.

This also allows us to slightly simplify the code by getting rid of the
ITEM_NAMESPACE_NODE_METHOD_LOCAL, which was only used to detect if we
should drop a reference on exit.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>