tables: use the sdt header when dumping FACS
The acpi_facs structure is not packed and therefore expects natural
alignment when used, which is not guaranteed since the header we're
dumping is created on the stack, and therefore has an alignment of 1.

Use the sdt header, which we already have defined above instead to
allow misaligned access in this case.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>