opregion: add a fast path to uacpi_reg_all_opregions
We don't need to lock anything nor jump inside a different function if
the space we're given doesn't require _REG calls. Do the check before
we do anything.

Also add this check to uninstall_address_space_handler (the final caller
of reg_or_unreg_all_opregions that didn't have an early check). With
that, we can now drop the check inside, which is now redundant.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>