kernel_api: calloc -> alloc_zeroed & make optional
calloc is an API that comes from the C land, so all C developers are
very familiar with it, however, people that do not come from the C land
have trouble understanding it and also have a tendency to forget to
zero the returned memory block.

Let's make this API a bit more user-friendly (aka does what it says on
the tin), while also making it optional so there's even less chance this
can go wrong.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>